requirements.txt
base36==0.1.1
imgui==2.0.0
nbtlib==2.0.4
numpy==2.3.5
pygame==2.6.1
pyglm==2.8.3
PyOpenGL==3.1.10
